refactor exceptions into specific models
