+++ src/main/java/seedu/address/model/group/exceptions/DuplicateGroupException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.group.exceptions;
 /**
  * Signals that the operation will result in duplicate Groups (Groups are considered duplicates if
  * they have the same identity).
