+++ src/test/java/seedu/address/model/group/UniqueGroupListTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.person.ActivityList;
 import seedu.address.model.person.Name;
 import seedu.address.model.person.PlaceList;
-import seedu.address.model.person.exceptions.DuplicateGroupException;
+import seedu.address.model.group.exceptions.DuplicateGroupException;
 
 public class UniqueGroupListTest {
